# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Arch Visualizer.",
        "title": "Arch Visualizer Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/v1",
    "paths": {
...
        "/static/cloud-config": {
            "get": {
                "description": "Get global cloud configurations like regions, instance types, etc.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "static"
                ],
                "summary": "Get cloud configurations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name (e.g. aws)",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/static/providers": {
            "get": {
                "description": "Get a list of all supported cloud providers (AWS, GCP, etc.)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "static"
                ],
                "summary": "List supported cloud providers",
                "responses": {
                    "200": {
                        "description": "List of providers",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/static/resource-models": {
            "get": {
                "description": "Get the JSON structure of resource output models grouped by service category with default mock data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "static"
                ],
                "summary": "Get resource output models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name (e.g. aws)",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/static/resource-types": {
            "get": {
                "description": "Get a list of all supported resource types, grouped by service category (e.g. Compute, Storage)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "static"
                ],
                "summary": "List resource types",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider name (e.g. aws)",
                        "name": "provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}